refactor : Auth Token 타입 구현
