+++ igoMoney/igoMoney/Source/UserClient/UserClient + live.swift
@@ -8,29 +8,6 @@ import Foundation
 
 import ComposableArchitecture
 
-struct AuthToken: Codable, Equatable {
-  let userID: Int
-  let accessToken: String
-  let expiresAt: String
-  let idToken: String
-  let refreshToken: String
-  let tokenType: String
-  let createdAt: Date = Date()
-  
-  var isExpired: Bool {
-    return createdAt.addingTimeInterval(TimeInterval(expiresAt) ?? 0) < Date().addingTimeInterval(86400)
-  }
-  
-  enum CodingKeys: String, CodingKey {
-    case userID = "userId"
-    case accessToken = "access_token"
-    case expiresAt = "expires_in"
-    case idToken = "id_token"
-    case refreshToken = "refresh_token"
-    case tokenType = "token_type"
-  }
-}
-
 extension UserClient {
   static var liveValue: UserClient = {
     @Dependency(\.apiClient) var apiClient
